requirements.txt
amazon-transcribe-streaming-sdk
sounddevice
groq
boto3
pygame
numpy 